refactor: Update import paths for user model across authentication routes
